Add drop tables as option default is drop all tables

+++ app/db/session.py
@@ -1,7 +1,6 @@
 import sqlalchemy as sa
 from sqlalchemy.orm import sessionmaker
-from sqlalchemy.ext.asyncio import create_async_engine
-from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
 from app.models.modelbase import SqlAlchemyBase
 import asyncio
 
@@ -19,50 +18,40 @@ def global_init_sqlite(db_file: str):
     conn_str = "sqlite:///" + db_file.strip()
     print(f"Connecting to DB with {conn_str}")
 
-    # echo=False to silience sql commands
     engine = sa.create_engine(conn_str, echo=True)
 
     factory = sessionmaker(bind=engine)
 
     # To sqlalchemy know all models to build
-    # from app.models.menu import Menu
-    # from app.models.submenu import Submenu
-    # from app.models.dish import Dish
     import app.models.__all_models
 
     SqlAlchemyBase.metadata.create_all(engine)
 
 
-def global_init_pg():
+def global_init_pg(drop: bool = False):
+    """Asyncio with postgresql"""
     global factory
     if factory:
         return
 
-    # if not db_file or not db_file.strip():
-    #     raise Exception("You must specify a db file.")
-    # conn_str = "postgresql://postgres:admin2255@localhost:5432/ylab_w1"
     conn_str = "postgresql+asyncpg://postgres:admin2255@localhost:5432/ylab_w1"
     print(f"Connecting to DB with {conn_str}")
 
-    # echo=False to silience sql commands
-    # engine = sa.create_engine(conn_str, echo=True)
     engine = create_async_engine(conn_str, future=True, echo=True)
 
     factory = sessionmaker(bind=engine, expire_on_commit=False, class_=AsyncSession)
 
     # To sqlalchemy know all models to build
-    # from app.models.menu import Menu
-    # from app.models.submenu import Submenu
-    # from app.models.dish import Dish
     import app.models.__all_models
 
     print(type(SqlAlchemyBase))
     print(type(SqlAlchemyBase.metadata))
-    # SqlAlchemyBase.metadata.create_all(engine)
 
-    async def init_models():
+    # Async with dropCreation
+    async def init_models(drop_all: bool):
         async with engine.begin() as conn:
-            await conn.run_sync(SqlAlchemyBase.metadata.drop_all)
+            if drop_all:
+                await conn.run_sync(SqlAlchemyBase.metadata.drop_all)
             await conn.run_sync(SqlAlchemyBase.metadata.create_all)
 
-    asyncio.run(init_models())
+    asyncio.run(init_models(drop))
